requirements.txt
asgiref==3.5.2
Django==3.2.16
django-rest-multiple-models==2.1.3
djangorestframework==3.14.0
drf-nested-routers==0.93.4
mysqlclient==2.1.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
